Update patch set 1
Patch Set 1:

I'd actually be curious to know if we can build our Exynos5420 HALs (picassowifi, mondrianwifi, lt03wifi etc.) from asp default repos (hardware/samsung_slsi/exynos5) which is used for manta right now.
Back when the Hardkernel guys released their Exynos Kitkat BSP and Ethan brought it in, it was newer than Manta's kitkat HAL stack. Now, not so much... But I'm not sure if those are applicable to Exynos5420, because Manta is 5250.

Patch-set: 1