Update patch set 3
Patch Set 3:

(1 comment)

I copied the whole function cause the sturct and the ioctl are different. This means I would have to add serveral #ifdef inside of the function which makes it harder to read. As the PRESET API is deprecated it is easier to remove it this way later.

Patch-set: 3